Added Progress bar for fund from other sources. Notifications changed to get top 15 messages

+++ src/main/java/org/codealpha/gmsservice/controllers/DisbursementsController.java
@@ -86,7 +86,7 @@ public class DisbursementsController {
                 List<Grant> grantsToReturn = new ArrayList<>();
                 if (ownerGrants != null && !ownerGrants.isEmpty()) {
                         for (Grant g : ownerGrants) {
-                                Long actualDisbursementAmount = disbursementService.getAcutalDisbursementAmountByGrant(g.getId());
+                                Double actualDisbursementAmount = disbursementService.getAcutalDisbursementAmountByGrant(g.getId());
                                 if (actualDisbursementAmount < g.getAmount()) {
                                        grantsToReturn.add(g);
                                  }
